requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-docx==0.8.11
werkzeug==2.3.7
gunicorn==21.2.0